Update userlogin.php
Changed to check if the result is empty.
Before, if you inputted an invalid username/password combination, an error would be returned when fetching the number of rows and checking if it equals to one. Ie. Notice: Trying to get property of non-object in C:\XAMPP\htdocs\304p3\userlogin.php on line 29

Now I believe it's the same behaviour, except the error won't be shown to the user.